Improve logging and rule_map performance.
Improve logging visual consistency; always prefix with 'make:' and put
context (file/lineno) after log level so things align better.
Store 'RuleMap' rules in a vector for storage and then use a `by_target`
hashmap to prevent having to store many clones of each rule (which is
the case if the hashmap is the primary storage).
Set release profile to use opt level 's', not 'z' so we don't skip loop
vectorization.